+++ springFinal/src/main/webapp/WEB-INF/views/student/student_join.jsp
